pyproject.toml
[project]
name = "rag-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "langchain>=1.2.15",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.12",
    "openai>=2.30.0",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
]
